DEBUG is now a macro for printing
To enable debug output, define CONFIG_DEBUG in the top level Makefile
(or uncomment) the example there. There is no garuntee of compatibility
of this new DEBUG usage with the old usage with ifdefs. While it will
enable correctly, the RELEASE define will no longer work as expected. I
will need to talk to John about depricating it since I'm not sure if
there are actually other uses for it besides auto disabling debug output.

TODO:
* Add some actual comments to the header so users know what to define\
  and how to use the library.
* Move over all the old usage of #ifdef DEBUG with printk to this new\
  style.
* Properly automagically include debug.h in all kernel c files and drivers.